+++ src/sentry/api/endpoints/integrations/sentry_apps/index.py
@@ -17,7 +17,7 @@ from sentry.auth.staff import is_active_staff
 from sentry.auth.superuser import is_active_superuser
 from sentry.constants import SentryAppStatus
 from sentry.models.integrations.sentry_app import SentryApp
-from sentry.sentry_apps.apps import SentryAppCreator
+from sentry.sentry_apps.logic import SentryAppCreator
 from sentry.users.services.user.service import user_service
 
 logger = logging.getLogger(__name__)
