+++ src/sentry/integrations/vercel/integration.py
@@ -27,11 +27,11 @@ from sentry.models.integrations.sentry_app_installation_token import SentryAppIn
 from sentry.organizations.services.organization import RpcOrganizationSummary
 from sentry.pipeline import NestedPipelineView
 from sentry.projects.services.project_key import project_key_service
+from sentry.sentry_apps.logic import SentryAppCreator
 from sentry.shared_integrations.exceptions import ApiError, IntegrationError
 from sentry.users.models.user import User
 from sentry.utils.http import absolute_uri
 
-from ...sentry_apps.apps import SentryAppCreator
 from .client import VercelClient
 
 logger = logging.getLogger("sentry.integrations.vercel")
