+++ src/sentry/sentry_apps/services/hook/impl.py
@@ -4,7 +4,7 @@ from django.db import router, transaction
 
 from sentry import deletions
 from sentry.models.servicehook import ServiceHook
-from sentry.sentry_apps.apps import expand_events
+from sentry.sentry_apps.logic import expand_events
 from sentry.sentry_apps.services.hook import HookService, RpcServiceHook
 from sentry.sentry_apps.services.hook.serial import serialize_service_hook
