+++ tests/sentry/api/endpoints/test_organization_sentry_app_installations.py
@@ -3,7 +3,7 @@ from django.test import override_settings
 from sentry.constants import SentryAppStatus
 from sentry.integrations.models.integration_feature import Feature
 from sentry.models.integrations.sentry_app_installation import SentryAppInstallation
-from sentry.sentry_apps.apps import SentryAppUpdater
+from sentry.sentry_apps.logic import SentryAppUpdater
 from sentry.slug.errors import DEFAULT_SLUG_ERROR_MESSAGE
 from sentry.testutils.cases import APITestCase
 from sentry.testutils.helpers.features import with_feature
